Writer Class
Moved the new function stack into the writer class, just like how the other modules are structured. I like the layout of using a class for these kinds of tasks, because there's some pieces of state that need to be mutated with each function call, and it's not always the most convenient to pass that state around between each function individually, so keeping it somewhere local where the functions can instead mutate on, seems a little nicer to understand for me. If you rely solely on return types for the function state to be passed around, it seems a bit more challenging to track where the state of the stack goes. So while this setup is a bit beefier, I like the more-easily configurable aspect to it.

So, in this state, it's simply just a wrapper class with methods that work together using their parameters and return types. After this stage, I'm going to move the state out of the function parameters, and into the class instance itself, that was each function can access the same state, but it doesn't need to be passed into it directly.