Method Flattening
Now that I broke up the Stringify implementation into methods on the writer class, I can de-abstract some of the function calls to be more specific to what some parts are doing, rather than running it through any uneccesary type checking and steps, since the type of the data being worked with at a given point is more known directly inside of each method. No need to figure out what kind of primitive it is, a second time.