Nullish Tag Type
Rather than using the `.indexOf()` kind of return type for unfound items, I'm moving the `getTagType()` invalid type return type to use `null` instead of `-1`, which is more like how calls like `querySelector()` work.

I'm changing to this method because I like the idea of being able to use the TypeScript `!` operator on the `getTagType()` return type, if I know that the tag being passed in as a check, is already verified to be a safe tag type. Essentially, if I've already filtered out any values that aren't valid tag primitive types, then asserting that the value is indeed a `TAG` value is easier to do with `getTagType()!` than with `getTagType() as TAG`, which isn't as explicit I think. You can more easily see what the first one is meaning to do.

I started trying to add inferred return type overloading working for `getTagType()`, but it's not turning out as straightforward initially, so I'm commiting these steps before trying to work on that :)