Tag Type Map
This works, fairly well! It's still a bit ugly to me, though.

This is the type matching provided by `getTagType()`, which should return the corresponding `TAG` for the given `Tag` value, and `null` if it's a different, non-NBT type. This is all for typing's sake, so it's mostly just for stronger type checking, and helping to make internal code smarter so I don't have to assert the value type, if the tag type of that value indeed is part of the `TAG` type, and not `null`. This doesn't appear to be working inside of the Write module though, so I think I either don't have it set up right yet, or don't quite know how to do what I'm trying to do. Either way, this is actually a very ergonomic way of doing it, and it was inspired by how the type checking for `ParentNode.querySelector()` and similarly-typed functions are typed, by using the `HTMLElementTagNameMap` interface, which pairs HTML tag string names with their corresponding `HTMLElement`-based classes. I essentially want to make the same exact behavior for this here, but instead of using strings for the parameter, to have a map from the tag value type, to the tag type ID, aka `{ [type: Tag]: TAG; }`, but that doesn't appear to be something you can do right now with type declarations.

The closest way that I found was to make a ternary operator check, which appears to essentially do the same thing. I had it working with this code here, then I also found this site, which appears to be using the same method I stumbled across. If someone made an article about doing it like this, then it at least looks like a possible way to go that may work! :D

https://www.javiercasas.com/articles/typescript-type-families

Not sure if I will keep this functionality over time, but I did want to document that I did have it working this time.