Parse Patterns & Tag Type Checks
Flattened more of the parse module by moving the RegExp SNBT string patterns into the function declarations directly.

I also removed my experimental type parameter narrowing for the `getTagType()` function, as I found that it helps to require stronger type handling from it's return values, since you want to use it to ensure that the type is what you think it is.

I think I should make an `isTag()` function, which will instead check for a specific type of tag that you provide it, and it can narrow the type to your `if` check scopes for you, so you don't have to use the `as StringTag` syntax, for example. So it would be something like `isTag(value,8)`, and it's return type would be narrowed with the return type narrowing syntax `function isTag<T extends Tag>(value: any): value is T`. The second parameter could be a specific type of tag that you would like to ensure the value is. I'll probably have to use function overloading to make that work :)

The moon is shinin' over the wall right now, and it looks really cool :)