NPX TSC
Super awesome!!! This is what I've been hoping to be able to do for a while now. Rather than using a 'devDependencies' install of a library you may need, turns out it's what npx is for, you can run any given library from npm with one command, and it will first look to see if you have it globally installed. And if not, it will cache a copy of it on your machine, and use that to run the command directly. I really like this, since I don't have to install TypeScript multiple times around. On my 16 GB Linuxbook, there's not as much extra room to have TypeScript installed 5 times, for all of my projects. This makes it so your commonly-used dependency toolchains only have to be saved in one place, and npm will scout those out for you and use them wherever they are! :D

Started looking into this when adding the `npm run dev` script for `tsc -w`, and I wasn't sure whether you had to use `npx` or anything special like that. So glad I stumbled across what npx is for! Ooooh!!!

This is going to make things like working on Electron apps much more performant, for the dev setup at least (maybe, just and idea :) ).

https://dev.to/azure/using-npx-and-npm-scripts-to-reduce-the-burden-of-developer-tools-57f9
https://stackoverflow.com/questions/50605219/difference-between-npx-and-npm
https://stackoverflow.com/questions/68719700/does-running-an-npm-package-with-npx-cause-devdependencies-to-get-installed

I'm also going to start using end-of-file new lines more often, as that seems to be the standard in most cases. npm adds it automatically to the `package-lock.json` after every `npm install` and I was manually removing it myself so it didn't mess up and Git changes. That's been super annoying, so I'm going to get myself to like end-of-file new lines so I don't have to worry about it anymore, haha. It seems to be the standard over in Linux/Unix land too, so I think it's smart to start getting used to it being the normal thing to expect.