SNBT Level Scoping + SNBT Stringed Object
Figured out the scoping setup for the SNBT stringify indentation level state. I was having trouble moving the level property out of the stringifier method calls, and onto the `SNBTWriter` object itself. Turns out I was just missing a few decrement calls when exiting a `ListTag` or `CompoundTag` level, and that was making the identation continue to expand out the further you went into the structure, rather than re-denting back to the number line again. Essentially, I needed the additional `this.#level -= 1;` calls after the `ListTag` and `CompoundTag` writing methods are run. I think I nearly had this working sometime last week, but I got it working for sure this time, after taking a look at what I was experimenting with again.

I added a type assertion for the root `parse()` call going from SNBT, as it was just the base `Tag` type before, and I know it will only ever be a `CompoundTag` for the root NBT tag. (I think I heard/read that Bedrock sometimes has a `ListTag` as the root element, but that's a separate feature I will look into, unrelated to this one. Gonna go make an issue/suggestion for myself about that right now, now that I mention it! ...opened: https://github.com/Offroaders123/NBTify/issues/23)

Moved the testing `NBTData` object for SNBT reading and writing into a separate module in the `./test/nbt` directory, since I plan to come back to it to check how the parsing is in the future, just like with the other NBT test files.