Parse Formatting
Doing some more work to be able to understand the base logic for Deno-Minecraft's SNBT parse functionality, as it's not as modular as I'd like it to be, and I don't quite understand how all of it works yet to rework it into my own thing. Doing bit by bit with what I know how to translate!