Build Script Learning
Found some more info about the build script searching that I was doing with the last commit, so I wanted to add links to new findings that I found after commiting it. These have to do with the CLI script setup, making your package able to be run as a command directly in the terminal!

https://stackoverflow.com/questions/10396305/npm-package-bin-script-for-windows
https://www.youtube.com/watch?v=dfTpFFZwazI

Didn't have anything fancy to commit just yet, but I did notice a little nice formatting that could be done to the Compression Streams API typings, so that was what I used to commit this message :) Found some more info with some other findings related to TS typings for Function Constructs, which is the syntax I am modifying in this commit. I was curious if other people use the short-spaced syntax for the `new()` method, or the long-spaced one, which I do think looks better. Was happy to see that it looks like the standard lib typings are likely short-spaced just because of formatting, and most people use the fancier formatting for their own code, making it look like `new ()`, which I like :)

https://stackoverflow.com/questions/13407036/how-does-interfaces-with-construct-signatures-work
https://stackoverflow.com/questions/66874130/how-to-properly-use-functions-construct-signatures-in-typescript