SNBT Compound Filtering
Added the unsupported entry filtering behavior for the SNBT writing process. This makes it work the same as the raw NBT writing process, where both will filter unsupported JavaScript primitives out of the data that you are passing in. This may seem to counteract with my goal of not changing the user's incoming data, but I think it's a bit different for these values, since they are just plain non-supportable, they don't have NBT equivalents, and it's how the JSON module deals with non-JSON serializable JavaScript primitives. Essentially, these are Functions, Symbols, `undefined`, and `null`. These will be silently skipped over during the write process, which does not yield them on the returned reading end.

I guess my rules for dealing with the user's data are specific to the supported primitives, rather than the unsupported ones. Each supported primtive has a specific use case, and it shouldn't be inconsistently used/managed by the library.

Forgot to mention this is the test data structure I was using for Class Compounds and List Item Type Coalescing testing a few weeks ago.