SNBT in Stable!
Publishing the last two months of updates to npm! It includes a lot of fixes and adjustments, with the headlining features being:

- `IntArrayTag` and `LongArrayTag` length reading and writing fix
- Deflate (zlib) compression support
- Class-Generated Compound Tags
- Less-Strict Write Handling
- List Item Type Coalescing (explored)
- SNBT read and write support
- SNBT writing indentation
- npx build scripts (removing dev dependencies!)

Pretty cool that I just turned 20, and the version of the package also happens to be 20! Didn't plan that. I update the number by the addition of new features/fixes along the development process, which is why it skips a few numbers sometimes.

I also didn't plan for it to be the same as the upcoming Minecraft update, that also wasn't planned! That's pretty neat too though, haha. Now my age is the same as the current Minecraft update, hadn't seen that before either. I guess that's the first time that will be the case, since Mojang also isn't the most straightforward with their versioning stategery, haha. Was trying to think if any other Minecraft updates also lined up with my age.

Remainder the Black Dog is awesome!!! So is Raider II; so happy that I bought Grace for Drowning, on Bandcamp Friday too! :O