EOF Strict Option
After using NBTify with my other world data parsing projects, I found that a few files weren't parsing correctly, even though they were opening as valid NBT in other programs, editors, and parsers. (I tried in VSCode's NBT editor extension, webNBT, and Prismarine-NBT. The first two opened them with no issue, while they both appeared to open the files with all of the data in them)

I looked into how Prismarine-NBT does their EOF checking, and it turns out that they only do that when you don't specify the file format of the NBT file. This means that Prismarine-NBT also fails to open these files when you don't specify the formats, but they open correctly when you do provide them. I do like that it still is able to open the file, I don't like that it makes it less strict in terms of possibly lost data though. I don't think many NBT files would be corrupt like that, but I still think I want that information to be forward facing first, in front of always being able to open anything from the file.

```ts
import { readFile } from "node:fs/promises";
import * as NBT from "prismarine-nbt";

const PLAYER_FILE_0 = "./P_280dfc7dac2f_00000001_knarF_520.dat";
const PLAYER_FILE_1 = "./N_280dfc7dac2f_100000001_.dat";

const data = await readFile(PLAYER_FILE_0);
console.log(data.length,"\n");

const result = (
  // await NBT.parse(data)
  await NBT.parse(data,"big")
);
console.log(result,"\n");

const recompile = NBT.writeUncompressed(result.parsed);
console.log(recompile,"\n");

console.log(Buffer.compare(data,recompile));
```

So to compromise and find the sweet spot in between, I'm adding a flag to disable EOF checks when reading any given file. This is what I will use to open files from Legacy Console Edition `GAMEDATA` file archives, as some NBT files from there appear to be sized to include extra bytes at the end of the file, which don't include any NBT data itself. At least, it looks like it's old data not needed anymore. For things like Dovetail, I will show an error to the user when it fails to open with the EOF checks (current expected behavior), and instead it will also provide them with the option to continue trying to open the file even though it's possible some data in the file may be lost when resaving again. It's most likely that the data at the end of the file is just junk bytes, but I can't be sure that their file may just not be something not supported by NBTify. The goal is to have all format coverage supported, but I'm not going to break users' files just to toat that.

I don't want to go with the Prismarine-NBT method which only does the checks when the format isn't specified, because that would be silently skipping the missed bytes at the end, and the library doesn't provide any notice to that happening. So user accessible or not, the library itself isn't keeping notice that some bytes at the end were missed. I think this is mostly a concern for me because I have tried opening random files with NBTify before, and they have succeeded to open, when they weren't actually even NBT files at all. This raises the likelyhood for false-positives, and that's not something I want to have to worry about. I'd rather that the issues be raised in the reading logic itself, then the API user can deal with those as necessary. The other options would be for the API user to need to manually check that the output NBT data tree is in the shape that they expect for their use case. I think that should be done either way (I'm talking about like TypeScript and Zod), but it shouldn't be required specifically to handle the output of NBTify itself.

The new `strict` option doesn't apply to finding TAG_End bytes when calling `NBTReader.prototype.#readTag` because there's no way to escape from an error there. If you encounter an unexpected tag type there, it might not be at the end of the file, it could be in the middle of it. So it would jsut finish with a tree only partially recursed into. At least, I think that's what would happen. I haven't had that error occur yet I think, so I will look into making a test file for that specific case when it occurs.

I realized that it would actually not be too hard to make dedicated broken NBT archives which the library could run unit tests on, and it would check if the appropriate errors occur for those instances and scenarios. I was already looking into unit tests for the different library formats, but also making specifically broken files would actually be another strong way to test if things are being caught properly.

(Reference to an issue for this opened a little while back, mentioning this to capture the discussion process #24)