TypeScript Interface Definition Generator
Wow, this is the 200th commit for this project! I can't believe how much I've put into it over just the past ~year or so. I'm really proud of this project, and it's really working out much better than I thought it would initially :)

Realized that I can automagically generate the conceptual TypeScript NBT interfaces I've been talking about, using my SNBT writing logic! I tried making a few edits to the current SNBT writing generator, and it worked out perfectly! I made a few more adjustments to make it nicer, and added some documentation, and then here we are.

This API isn't totally done yet, so it's not a stable feature yet. I think it will work stable-y though, so there shouldn't be too much to worry about with it providing broken type definitions.

There are some limitations to how this works. Since it's essentially generating types for the NBT tree you give it, it can only know about the properties and values that are actually inside of that tree when you give it to the generator. There might be optional NBT properties or things like that, which you may have to manually add yourself because there wouldn't be a way for the generator to know to define them, without them being there in the tree itself.

This is of similar interest to the difference between `ByteTag` and `BooleanTag` values. Since NBT only really has `ByteTag`s, and not `BooleanTag`s, it's up to you to modify tags read as `ByteTag` values and whether they can be scoped down to the more-specific, `BooleanTag` representation. This will allow for better type narrowing and values which the user can add there, as you wouldn't want `CenterMapsToOrigin` to have a value of `8`. Not sure what could happen there "probably not much", but you don't want to allow errors like that into the editing process when they don't have to be there. Narrowing the type also allows the user to use boolean literals like `true` and `false`, which NBTify will coalesce back down to the actual `ByteTag` representation again. It allows it to be one less step of needing to write `new Byte(0)` or `new Byte(1)`, which is nice for the user :). And, in reality, these are just types, so it will actually still work either way. But, you don't want to throw errors to the user for values that are actually working as intended (providing `true` to a `ByteTag` will work fine in terms of NBTify being able to handle that. NBTify will write whatever you throw at it down to NBT, that's the nature of JavaScript, since there are no run-time types themselves. I chose with NBTify to leverage types to provide the checking for the user, instead of run-time errors).

With all of the notes provided above, here's a result of the types for a Bedrock `level.dat` created by the generator (I made some edits to make it work in actual TypeScript):

```ts
import { BooleanTag, IntTag, LongTag, FloatTag, StringTag } from "nbtify";

export interface BedrockLevel {
  BiomeOverride: StringTag;
  CenterMapsToOrigin: BooleanTag;
  ConfirmedPlatformLockedContent: BooleanTag;
  Difficulty: IntTag;
  FlatWorldLayers: StringTag;
  ForceGameType: BooleanTag;
  GameType: IntTag;
  Generator: IntTag;
  InventoryVersion: StringTag;
  LANBroadcast: BooleanTag;
  LANBroadcastIntent: BooleanTag;
  LastPlayed: LongTag;
  LevelName: StringTag;
  LimitedWorldOriginX: IntTag;
  LimitedWorldOriginY: IntTag;
  LimitedWorldOriginZ: IntTag;
  MinimumCompatibleClientVersion: [IntTag, IntTag, IntTag, IntTag, IntTag];
  MultiplayerGame: BooleanTag;
  MultiplayerGameIntent: BooleanTag;
  NetherScale: IntTag;
  NetworkVersion: IntTag;
  Platform: IntTag;
  PlatformBroadcastIntent: IntTag;
  RandomSeed: LongTag;
  SpawnV1Villagers: BooleanTag;
  SpawnX: IntTag;
  SpawnY: IntTag;
  SpawnZ: IntTag;
  StorageVersion: IntTag;
  Time: LongTag;
  WorldVersion: IntTag;
  XBLBroadcastIntent: IntTag;
  abilities: {
    attackmobs: BooleanTag;
    attackplayers: BooleanTag;
    build: BooleanTag;
    doorsandswitches: BooleanTag;
    flySpeed: FloatTag;
    flying: BooleanTag;
    instabuild: BooleanTag;
    invulnerable: BooleanTag;
    lightning: BooleanTag;
    mayfly: BooleanTag;
    mine: BooleanTag;
    op: BooleanTag;
    opencontainers: BooleanTag;
    permissionsLevel: IntTag;
    playerPermissionsLevel: IntTag;
    teleport: BooleanTag;
    walkSpeed: FloatTag;
  };
  baseGameVersion: StringTag;
  bonusChestEnabled: BooleanTag;
  bonusChestSpawned: BooleanTag;
  commandblockoutput: BooleanTag;
  commandblocksenabled: BooleanTag;
  commandsEnabled: BooleanTag;
  currentTick: LongTag;
  dodaylightcycle: BooleanTag;
  doentitydrops: BooleanTag;
  dofiretick: BooleanTag;
  doimmediaterespawn: BooleanTag;
  doinsomnia: BooleanTag;
  domobloot: BooleanTag;
  domobspawning: BooleanTag;
  dotiledrops: BooleanTag;
  doweathercycle: BooleanTag;
  drowningdamage: BooleanTag;
  eduOffer: IntTag;
  educationFeaturesEnabled: BooleanTag;
  experiments: {
    experiments_ever_used: BooleanTag;
    saved_with_toggled_experiments: BooleanTag;
  };
  falldamage: BooleanTag;
  firedamage: BooleanTag;
  freezedamage: BooleanTag;
  functioncommandlimit: IntTag;
  hasBeenLoadedInCreative: BooleanTag;
  hasLockedBehaviorPack: BooleanTag;
  hasLockedResourcePack: BooleanTag;
  immutableWorld: BooleanTag;
  isFromLockedTemplate: BooleanTag;
  isFromWorldTemplate: BooleanTag;
  isSingleUseWorld: BooleanTag;
  isWorldTemplateOptionLocked: BooleanTag;
  keepinventory: BooleanTag;
  lastOpenedWithVersion: [IntTag, IntTag, IntTag, IntTag, IntTag];
  lightningLevel: FloatTag;
  lightningTime: IntTag;
  limitedWorldDepth: IntTag;
  limitedWorldWidth: IntTag;
  maxcommandchainlength: IntTag;
  mobgriefing: BooleanTag;
  naturalregeneration: BooleanTag;
  prid: StringTag;
  pvp: BooleanTag;
  rainLevel: FloatTag;
  rainTime: IntTag;
  randomtickspeed: IntTag;
  requiresCopiedPackRemovalCheck: BooleanTag;
  respawnblocksexplode: BooleanTag;
  sendcommandfeedback: BooleanTag;
  serverChunkTickRange: IntTag;
  showbordereffect: BooleanTag;
  showcoordinates: BooleanTag;
  showdeathmessages: BooleanTag;
  showtags: BooleanTag;
  spawnMobs: BooleanTag;
  spawnradius: IntTag;
  startWithMapEnabled: BooleanTag;
  texturePacksRequired: BooleanTag;
  tntexplodes: BooleanTag;
  useMsaGamertagsOnly: BooleanTag;
  worldStartCount: LongTag;
  world_policies: {};
}
```

Note how the `world_policies` key is empty. I haven't looked into the docs for it yet, but I imagine it likely holds optional keys and values there, which would also be nice to have the types of. So, after using the generator to build the types scaffolding, it will go a long way if you go the extra mile to adjust the types to your own findings so it can fully reflect the NBT trees you are working with!

Oh yeah, and one more part. Adding JSDoc comments to describe what each key does would also be a great help for working with the data. Combining this with the descriptions provided by the Minecraft Wiki, for example, would be very powerful to have right when editing your NBT data. (I could do this for this definition file with the docs here: https://minecraft.fandom.com/wiki/Bedrock_Edition_level_format#level.dat)

One other part I forgot to mention: It would also be a great detail to narrow the values provided by a certain type itself. I'll have to dig more into the docs yet again to find these. But, for example, the `Difficulty` value probably should be represented more narrowly than just `IntTag`. On the Wiki, you can see that it mentions that it's values are `0 is Peaceful, 3 is Hard.`, so it could instead be represented with the type generic with the values accepted values themselves, `IntTag<0 | 1 | 2 | 3>`. Then you will get type errors when trying to set the difficulty to `56`, that sounds a bit too challenging XD (and, the fact that it's not real).