Read-Only Data Properties
I want to make the properties present on `NBTData` objects read-only, but I still want the class to allow itself to be extendable. Using `Object.freeze()` isn't very helpful for anyone that's hoping to do something custom of their own with the `NBTData` class.

I like the functionality for how this works, but I don't like the logging for it. In Node, you can't see the values for any of the getters, and in the browser, you can see both the private properties and their values (not that big of a deal, just a bit bloated), and you can see the getters for them right below. It won't render their values unfortunately, though. Unless you call the getters yourself by clicking on them.

This led me to the experimentation in the next commit, which has to do with custom inspect styles in Node. I found a lot, and it led to a train of discoveries after that too!