Demo Cleanup
Learned a lot the last few commits! Now I think I have more of a clearer idea again as to how the library should go from here, as to what's the strongest API for users. I think I want to continue sticking with using primitive wrapper objects as little as possible.

Going to start fresh on the required primitive wrapper custom logging logic, and try some other things with that now that I know how it works.

Listening to 'You Must Be This Tall', now :P

Gonna go eat something soon, gettin' hungerey.

Oh yeah, the other part of this commit, kinda forgot haha. Realized that I like having separate imports for types, it looks nice to have a specific import with only types in it, makes it a little easier to discern everything that you're adding to the module, and what it's used for. I also read up on some of the changes for TypeScript 5, which had a bunch of great things in there! @satisfies for JSDoc is killer. There's all kinds of neat things in there like that. Can't wait to stumble across another random use-case when another new language feature fill the missing void when I run into another issue.

https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-0.html#--verbatimmodulesyntax