Node Inspection + Compatibility Table
Added proper Node inspection methods to the NBTify primitive wrapper objects! Now it looks great in the console by default.

Was talking with some fellow Minecraft Manipulators about the new changes, and we were talking about how NBT is fairly close to JSON. Mentioned that's what this library is trying to do (as much as can be done), and I thought I'd make a compatibility chart to make note of all the different primitives and type overlaps between the different languages and formats. I remembered I'll have to add the remaining JavaScript primitives, and also the SNBT types too. This is great for laying things out! I had this kind of model in my head for this, for a while now. So, it's good to have a nice visual way of explaining that goal of lining up parts of the different mediums.