Node Inspect Visibility + Definition Name
Did some testing with the current code in the browser. I'm glad I checked it, because you could see the `nodejs.util.inspect.custom` method right in the logging, and it didn't look very good. My first workaround to hide it was to add it with `Object.defineProperty()` and to make it non-enumerable as a regular property, but that both didn't work, and it caused the whole thing to crash. Arg, so I moved on to another idea, which works great! Defining the property with a getter, then returning the actual function from the getter both hides the property (I think it adds the getter to the primitive wrapper's prototype), and it no longer crashes, lol. Perfect! It's still a bit verbose, which I don't quite like, but it's a nice improvement in the terminal, so I'm keeping it for the meantime.

Wanted to try out the definition generator in the browser too, just to see how it would work. It was great! The only problem was that I shouldn't have made the interface name an optional option for the function call, so I changed it to be mandatory now.