Error Formatting (Reverted)
Reverted these first changes before actually committing, turns out my knowledge about error formatting was actually the opposite, haha. Looks like neither the browser or Node has periods at the end of their error messages.

Realized that most browser errors also have a period at the end of the error message. Not that big of a thing, but a nice detail to help with the readability of things.

Also formatted the `nodejs.util.inspect.custom` usage in the Primitives module.