Read-only NBTData Properties
Figured out how to instead implement the NBTData class's read-only nature. This still allows for custom object changes if you know what you are trying to do, unlike when the read-only properties were established using `Object.freeze()`.

Discovered a lot about how getters, setters, readability, and things like that work with the prototype chain. There was a lot to learn, and I got a lot of helpful links to try figuring this out a bit more. Essentially, I want to make `NBTData` look and feel just like the properties on classes like `File`, `Blob`, and `ImageData` do. It's not quite built in the same way yet I think, but this new implementation's behavior is the most similar to it so far. I think the way to make it work is to define enumerable getters for the properties on the `NBTData` prototype, which can be visible as properties from the regular classes themselves, which inherit the prototype.

The closest I got code-wise to making this work, unfortunately isn't represented the same in the browser DevTools, because it shows the getter `(...)` instead of the currently calculated value. I just found one more SO post that seems to be covering this issue; I didn't find any solod solutions there as to what can make a getter not 'lazy'. It sonded like enabling Strict Mode for it might help, but I couldn't find any other information about that anywhere else.

https://stackoverflow.com/questions/29772403/cant-enumerate-getters-setters-properties
https://stackoverflow.com/questions/34517538/setting-an-es6-class-getter-to-enumerable
https://github.com/nodejs/node/blob/7bc0e6a4e79684a98cb64c61452af00e20f590e7/lib/internal/file.js
https://github.com/nodejs/node/blob/7bc0e6a4e79684a98cb64c61452af00e20f590e7/lib/internal/blob.js
https://stackoverflow.com/questions/54783604/what-is-happening-in-the-console-when-i-look-at-getter-properties-on-dom-objects
https://dev.to/mustapha/a-deep-dive-into-javascript-object-properties-598h
https://developer.chrome.com/blog/new-in-devtools-101/#private-props
https://github.com/nodejs/node/issues/27404
https://stackoverflow.com/questions/49710825/implement-smart-self-overwriting-lazy-getters-in-strict-mode