NBTData Generics
Added support for using generics types with the NBTData class, and it's surrounding functional APIs!

Now if you know the shape of the NBT data going into the read function, you can pass that type into the read function generic parameter. This makes it so it can be a little easier to manage the types for the NBT structures going in and out of the library.

This is better than just plain type assertions on top of the `NBTData.data` property, because now the `NBTData` class type itself can hold both the NBT file metadata itself, and the shape of the data that it holds. I'm planning on using this to describe the full NBT file data structure and file metadata for world save data entries, so then you know things like the endian format, compression type, and what the actual NBT in the file is. So `NBTData` can be a full representation of the entire NBT file and how to make it with NBTify! This will work great once I eventually need to make NBT-generating classes which can make defaults for the these files when necessary. The class will simply implement the file's `NBTData` interface shape (including the NBT structure, file metadata), then the class will be fully type-safe with all of the information needed to make a fully symmetrical data tree to that of what's generated by the vanilla game :)

Currently, my use of validating the NBT data trees against the `CompoundTag` interface is currently typed too-strictly for what I'm trying to use it for, and it's forcing the interfaces and classes to require establishing index signatures, which are partially eliminating what I want the interface definitions to enable in the first place. I don't want my interfaces to allow the access and additions of unspecified properties, I only want to check that the values I am defining are compatible with the NBT primitive types.

https://github.com/microsoft/TypeScript/issues/52222#issuecomment-1544785199
https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-9.html

I created a new issue (#28) under NBTify to track the development of this typing issue in terms of how NBTify is handling this problem.