Reverting NBT Primitive Type Narrowing
Removing the API-safe `CompoundTag` type checking for the library entry points, in the meantime. Without the index signature key value checking behavior that I need to accomplish the value type checking, there isn't another way to create API designs that will make the compiler happy, unfortunately. So this means I'm going to make all objects acceptable into NBTify, on the type level at least. Since I can't use type checking to validate the types for the shapes I'm passing in, I will just have to allow them in, and make sure that the shapes have appropriate value types.

While working on this revert, I looked into the `object` type, and realized it's kind of similar in a way as to what I'm trying to do with my primitive interface type checkers (`ListTag` and `CompoundTag`).

> Introduction to TypeScript `object` type
> The TypeScript `object` type represents all values that are not in primitive types.
>
> The following are primitive types in TypeScript:
>
> `number`
> `bigint`
> `string`
> `boolean`
> `null`
> `undefined`
> `symbol`

For my object definitions, I'm trying to define an object with keys that are only of a select few types. For the `object` type, it represents all JavaScript types that aren't a primitive type. So with this change, NBTify simply only checks if your `CompoundTag` is none of the primitive values mentioned above, which is better than just using `any` at least. It would be better if the check could also specify that your `CompoundTag` object cannot have things like `Function`s, `symbol`s, `undefined`, `null`, things like that. I think that's my main reason for wanting to add parameter type checking for your `CompoundTag` use.

https://www.typescripttutorial.net/typescript-tutorial/typescript-object-type/

https://github.com/microsoft/TypeScript/issues/52222
(#28)