Internal Type Insecurities
While the types for the tags are nearly perfect (of course, they can't actually really be perfect), I realize that the `ListTag` and `CompoundTag` types aren't typed strong enough, out of everything in the library. I'm going to work on refactoring things as a whole, there's starting to be more things I see being able to make improvements to, because I can move logic for things both into dedicated functions, and to make use of stronger type checking which will ensure that values are actually what they are when they are being accessed. Right now, I'm not quite doing the full checking on the shapes and content of each tag, and that could cause more problems that could be uncaught, even with the types I currently have for things.

I'm going to further strengthen the checking on each value and structure individually, rather than asserting it through error catching at the entry points to each part of the API. So, essentially I want to make more use of internally representing values as unknown, and only having type definitions on the outside, then type checking within the library itself. What that means, the users will see the intended types of the library from the outside, but the library will instead start treating all values as `unknown`, since I want to intentionally find out their types, and remove any unsafe data (things like the unsupported types `Function`, `symbol`, `null`, `undefined` for example). I don't want to do this magically, I want to start purposefully managing these checks using the internal types within the library itself. Right now I just have to remember to add these checks in myself, wherever I know they haven't been sanitized yet. So I think this in part has to do with my now understanding of the uncertainty behind not having proper API type checks within the library, which will check if the incoming `CompoundTag` structures are valid, including valid properties and such. And I think this means it will (un?)realistically allow for the API to have `unknown` entry point values, and the library should be still type safe internally, where to the point it will not throw errors, because the proper checks are in place to actually be able to deduce the types of the NBT data object structures themselves, not only by the API types themselves.

This is just an idea as of now, so I may not end up going through making everything this solidly strict. I think that would be very good in terms of strictness and safety, but it might be redundant in a few (or, a lot) of places. So I may just implement these ideas in places of the library where it tends to be less safe or forgiving than others (`CompoundTag` entries, `ListTag` items).

I want to move the library into a more functionally-driven nature I think now too, driven by modules and smaller functions, instead of logic-obscuring classes which hide implementation details away. While I do like that structure of defining things (at least it was pretty nice at first), I am starting to see more problems where I want to use similar logic in other places, and I can't apply those in other places outside of that one use case. One big example was the SNBT to type definition generator logic. Both are very similar in what they do, and parse the NBT data tree, but they aren't similar enough to be a single class, and I don't want to use inheritance to find the similarities between the two. Sounds veery ugly. As a single class, which does it's own things though, I guess I do like that model. Maybe if it's just for simpler, more standalone things it can be ok. I think working on my other world data projects, I found that using functions to pass data between each other is more dynamic, in the fact that you don't have to relate two things together in order to be able to use them with each other. You just pass the value from one call into the next. You don't have to use encapsulation to pool all of the state into a singular place, instead you can pipe it into different canals that go to different places, and eventually go to the finish line. The rivers can at times cross paths too, and use the same roadway (sorry for these horrible analogies today), and break off again, using a different function to modify it's output just a little bit different than it's neighbor Angela, who simply uses the same river (function) to get from the starting line to the finish line. (not sure who Angela is, just dove into making the analogy worse)